# server/openapi.yaml
openapi: 3.0.3
info:
  title: CEDEAR Advisor API
  version: 2.0.0
  description: |
    API para análisis, scoring y gestión de portfolios de CEDEARs.
    Incluye fallback multi-proveedor (Yahoo Finance, Finnhub, Stooq),
    motor de riesgo, backtests y asesoramiento con IA (Claude).
servers:
  - url: http://localhost:3001/api
    description: Local development
paths:
...
  /ranking:
    get:
      summary: Ranking cuantitativo de CEDEARs
      parameters:
        - in: query
          name: top
          schema: { type: integer, default: 20 }
        - in: query
          name: profile
          schema: { type: string, enum: [conservative, moderate, aggressive], default: moderate }
        - in: query
          name: sectors
          schema: { type: string }
        - in: query
          name: minScore
          schema: { type: integer, default: 55 }
        - in: query
          name: showDiversified
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Ranking con scores técnicos y fundamentales
  /cedear/{ticker}:
    get:
      summary: Datos completos de un CEDEAR
      parameters:
        - in: path
          name: ticker
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Quote, financials, analysis y signals
        "404":
          description: Ticker no encontrado
  /history/{ticker}:
    get:
      summary: Historial de precios
      parameters:
        - in: path
          name: ticker
          required: true
          schema: { type: string }
        - in: query
          name: range
          schema: { type: string, default: "1y" }
      responses:
        "200":
          description: Array de precios históricos
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT